fix: fixed several recursive dependencies
