+++ src/query-builder/param-parser.ts
@@ -1,11 +1,12 @@
 import { types } from 'util';
-import { RethinkDBErrorType } from '..';
+import { RethinkDBErrorType } from '../types';
 import { RethinkDBError } from '../error/error';
 import { TermJson } from '../internal-types';
 import { TermType } from '../proto/enums';
 import { globals } from './globals';
 import { isQuery, toQuery } from './query';
 import { isFunction, isObject } from '../util';
+import { hasImplicitVar } from './has-implicit-var';
 
 export function parseParam(
   param: any,
@@ -122,23 +123,6 @@ export function parseOptarg(obj?: any) {
     }),
     {},
   );
-}
-
-export function hasImplicitVar(term: TermJson | undefined): boolean {
-  if (!Array.isArray(term)) {
-    if (term !== null && typeof term === 'object') {
-      return Object.values(term).some((value) => hasImplicitVar(value));
-    }
-    return false;
-  }
-  if (term[0] === TermType.IMPLICIT_VAR) {
-    return true;
-  }
-  const termParam = term[1];
-  if (termParam) {
-    return termParam.some((value) => hasImplicitVar(value));
-  }
-  return false;
 }
 
 function camelToSnake(name: string) {
